Async download
Progress function added, and WinInet calls are now async.

The Async API (+ docs + example) for WinInet are truly dreadful. It's *really* easy to use it wrong.

This is my current understanding:

When you call InternetReadFile(), you might get 0 back (error), and
GetLastError() == ERROR_IO_PENDING, which isn't an error, just an indication that
it's not yet finished, and you have to wait for the INTERNET_STATUS_REQUEST_COMPLETED
callback. When that comes, the data + size are filled in, and you can continue.

We always wait on the callback event AND the cancellation token, so we can
aways cancel the download.